improve plugin links + luckperms link

+++ api/src/main/java/eu/darkcube/system/link/Link.java
@@ -7,24 +7,50 @@
 
 package eu.darkcube.system.link;
 
-import java.io.Closeable;
 import java.util.concurrent.atomic.AtomicBoolean;
 
-public abstract class Link implements Closeable {
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+public abstract class Link {
+
+    protected final Logger logger = LoggerFactory.getLogger(getClass());
 
     private final AtomicBoolean enabled = new AtomicBoolean(false);
+    private final AtomicBoolean link = new AtomicBoolean(true);
+    private final AtomicBoolean linked = new AtomicBoolean(false);
 
     public Link() {
     }
 
-    protected abstract void link() throws Throwable;
-
-    public final void enable() {
-        if (enabled.compareAndSet(false, true)) onEnable();
+    boolean isLinked() {
+        return linked.get();
     }
 
-    public final void disable() {
-        if (enabled.compareAndSet(true, false)) onDisable();
+    private void link0() throws Throwable {
+        if (link.compareAndSet(true, false)) {
+            link();
+            logger.info("Linked link {}", getClass().getSimpleName());
+            linked.set(true);
+        }
+    }
+
+    protected abstract void link() throws Throwable;
+
+    private void enable() {
+        if (!linked.get()) return;
+        if (enabled.compareAndSet(false, true)) {
+            onEnable();
+            logger.info("Enabled link {}", getClass().getSimpleName());
+        }
+    }
+
+    private void disable() {
+        if (!linked.get()) return;
+        if (enabled.compareAndSet(true, false)) {
+            logger.info("Disabling link {}", getClass().getSimpleName());
+            onDisable();
+        }
     }
 
     protected void onEnable() {
@@ -33,10 +59,26 @@ public abstract class Link implements Closeable {
     protected void onDisable() {
     }
 
+    private void unlink0() {
+        if (linked.compareAndSet(true, false)) {
+            logger.info("Unlinking link {}", getClass().getSimpleName());
+            unlink();
+        }
+    }
+
     protected abstract void unlink();
 
-    @Override
-    public final void close() {
-        unlink();
+    protected final void disableAndUnlink() {
+        disable();
+        unlink0();
+    }
+
+    protected final void tryLinkAndEnable() {
+        try {
+            link0();
+            enable();
+        } catch (Throwable t) {
+            logger.warn("Failed to enable link {}", getClass().getSimpleName(), t);
+        }
     }
 }
